Move adding default text to chat out of domain and into controller This is a client concern and not a domain concern
